+++ src/js/content/cellar/tiles/armory.js
@@ -5,24 +5,54 @@ content.cellar.tiles.armory = content.cellar.tiles.invent({
   uniquePerRun: true,
   weight: 1,
   defaultState: {
-    entered: false,
+    uses: 0,
   },
-  onEnterEffects: function () {
-    if (this.state.entered) {
-      return
+  canInteractMore: function () {
+    return this.state.uses < this.calculateMaxUses()
+  },
+  getDialogs: () => [
+    {
+      title: `It's a stockpile.`,
+      description: `Ghastly silhouettes of <em>earthen weapons</em> possess the manifold gaps of its frames. Its vesigial provisions gather rust and collect dust the longer that they age.`,
+      actions: [
+        {label: 'Adore the sword'},
+        {label: 'Fancy the shield'},
+        {label: 'Covet the amulet'},
+        {label: 'Stick with your fists'},
+      ],
+    },
+    {
+      tutorial: true,
+      title: `<span class="u-highlight">[Tutorial]</span> <span class="u-screenReader">for</span> The armory:`,
+      description: () => ({
+        gamepad: `${app.settings.computed.inputHold ? 'Hold' : 'Press'} any <kbd>Face Button</kbd>`,
+        keyboard: `${app.settings.computed.inputHold ? 'Hold' : 'Press'} <kbd>Enter</kbd> or <kbd>Spacebar</kbd>`,
+        mouse: `${app.settings.computed.inputHold ? 'Click and hold' : 'Click'} the <kbd>Interact Button</kbd>`,
+        touch: `${app.settings.computed.inputHold ? 'Tap and hold' : 'Tap'} the <kbd>Interact Button</kbd>`,
+      }[app.tutorial.getInputPreference()]) + ` to equip additional protection.`,
+    },
+  ],
+  getEffects: function () {
+    const effects = [
+      ...this.effectsUnique,
+    ]
+
+    if (this.state.uses >= this.calculateMaxUses()) {
+      effects.unshift({
+        attribute: {
+          label: `Empty shelves`,
+          modifiers: [],
+        },
+      })
     }
 
-    content.cellar.barrier.add(3)
-    content.audio.barrierChange.trigger({isUp: true})
-
-    this.effectsOnEnter.push({
-      attribute: {
-        label: 'Greed increased',
-        modifiers: ['barrier'],
-      },
+    return effects
+  },
+  getInteractLabelMore: () => 'Interact',
+  onInteractMore: function () {
+    content.location.emit('cellar-armory', {
+      tile: this,
     })
-
-    this.state.entered = true
   },
   alterParticle: function (particle) {
     const radius = 10,
@@ -46,11 +76,36 @@ content.cellar.tiles.armory = content.cellar.tiles.invent({
     particle.target.z += z * 2
 
     if (z > 0 && z < 1 && engine.fn.between(dy, 2.5, 7.5)) {
-      const time = content.time.value()
+      const time = content.time.value(),
+        value = 1 - engine.fn.clamp(this.state.uses / this.calculateMaxUses())
 
-      particle.target.h = Math.sin(engine.const.tau * time / 60 * particle.twinkleFrequencies[1])
-      particle.target.s = 3/4 + (1/4 * Math.sin(engine.const.tau * time * particle.twinkleFrequencies[2]))
-      particle.target.v = 3/4 + (1/4 * Math.sin(engine.const.tau * time * particle.twinkleFrequencies[0]))
+      particle.target.h = engine.fn.lerpExp(
+        -25/360,
+        Math.sin(engine.const.tau * time / 60 * particle.twinkleFrequencies[1]),
+        value,
+        1.5,
+      )
+
+      particle.target.s = engine.fn.lerpExp(
+        0.25,
+        3/4 + (1/4 * Math.sin(engine.const.tau * time * particle.twinkleFrequencies[2])),
+        value,
+        1.5,
+      )
+
+      particle.target.v = engine.fn.lerpExp(
+        0.25,
+        3/4 + (1/4 * Math.sin(engine.const.tau * time * particle.twinkleFrequencies[0])),
+        value,
+        1.5,
+      )
     }
+  },
+  // Interactions
+  calculateMaxUses: function () {
+    return Math.max(3, Math.abs(this.z))
+  },
+  incrementUses: function () {
+    this.state.uses += 1
   },
 }, content.cellar.tiles.baseUnique)